# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for using Swagger with Fiber.",
        "title": "Fiber Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/schools": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schools"
                ],
                "summary": "List all schools",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.HTTPSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.School"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schools"
                ],
                "summary": "Register a new school",
                "parameters": [
                    {
                        "description": "School information",
                        "name": "school",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.School"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.HTTPSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.School"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/schools/packages": {
            "post": {
                "tags": [
                    "schools"
                ],
                "summary": "Create application package",
                "parameters": [
                    {
                        "description": "Package details",
                        "name": "package",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AppPackage"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPSuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/schools/{id}": {
            "get": {
                "tags": [
                    "schools"
                ],
                "summary": "Get school by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "School ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.HTTPSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.School"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "schools"
                ],
                "summary": "Update school details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "School ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "School information",
                        "name": "school",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.School"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.HTTPSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.School"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/schools/{id}/license": {
            "post": {
                "tags": [
                    "schools"
                ],
                "summary": "Assign license to school",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "School ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Package Code",
                        "name": "package_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.HTTPSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.SchoolLicense"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/settings": {
            "get": {
                "description": "Get settings, optionally filtered by group_name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group Name (e.g. umum, backup, integrasi)",
                        "name": "group",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Settings retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.HTTPSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.SettingResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve settings",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update or insert multiple settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update settings",
                "parameters": [
                    {
                        "description": "Settings update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.SettingUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Settings updated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update settings",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.AppPackage": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_students": {
                    "type": "integer"
                },
                "max_teachers": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price_monthly": {
                    "type": "number"
                },
                "price_yearly": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "entity.LicenseStatus": {
            "type": "string",
            "enum": [
                "active",
                "expired",
                "suspended"
            ],
            "x-enum-varnames": [
                "LicenseActive",
                "LicenseExpired",
                "LicenseSuspended"
            ]
        },
...
        "entity.School": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "accreditation": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "npsn": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "principal_name": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "school_level": {
                    "$ref": "#/definitions/entity.SchoolLevel"
                },
                "status": {
                    "$ref": "#/definitions/entity.SchoolStatus"
                },
                "subscription_end_date": {
                    "type": "string"
                },
                "subscription_start_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "entity.SchoolLevel": {
            "type": "string",
            "enum": [
                "SD",
                "SMP",
                "SMA"
            ],
            "x-enum-varnames": [
                "LevelSD",
                "LevelSMP",
                "LevelSMA"
            ]
        },
        "entity.SchoolLicense": {
            "type": "object",
            "properties": {
                "app_package_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_key": {
                    "type": "string"
                },
                "school_id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entity.LicenseStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "entity.SchoolStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "suspended"
            ],
            "x-enum-varnames": [
                "SchoolActive",
                "SchoolInactive",
                "SchoolSuspended"
            ]
        },
...
        "entity.SettingItem": {
            "type": "object",
            "required": [
                "group_name",
                "setting_key"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "group_name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "integer"
                },
                "setting_key": {
                    "type": "string"
                },
                "setting_value": {
                    "type": "string"
                }
            }
        },
        "entity.SettingResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "group_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "school_id": {
                    "type": "integer"
                },
                "setting_key": {
                    "type": "string"
                },
                "setting_value": {
                    "type": "string"
                }
            }
        },
        "entity.SettingUpdateRequest": {
            "type": "object",
            "required": [
                "settings"
            ],
            "properties": {
                "settings": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/entity.SettingItem"
                    }
                }
            }
        },
...
        "response.HTTPErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.JSONError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.HTTPSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.JSONError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}